Build failure due to .so misisng

+++ source/core/wifi_ctrl_queue_handlers.c
@@ -2864,10 +2864,10 @@ int get_neighbor_scan_results(void *arg)
 
 void process_acs_keep_out_channels_event(const char* data)
 {
-    int numOfRadios = (int)getNumberRadios();
+    unsigned int numOfRadios = getNumberRadios();
     wifi_radio_operationParam_t *radio_oper = NULL;
     decode_acs_keep_out_json(data,numOfRadios);
-    for(int i=0;i<numOfRadios;i++)
+    for(unsigned int i=0;i<numOfRadios;i++)
     {
         radio_oper = (wifi_radio_operationParam_t *)get_wifidb_radio_map(i);
         if(radio_oper)
